+++ routes/naviera.js
@@ -72,7 +72,7 @@ app.post('/', mdAutenticacion.verificaToken, (req, res) => {
         rfc: body.rfc,
         razonSocial: body.razonSocial,
         nombreComercial: body.nombreComercial,
-        calle: body.calle ,
+        calle: body.calle,
         noExterior: body.noExterior,
         noInterior: body.noInterior,
         colonia: body.colonia,
@@ -89,12 +89,12 @@ app.post('/', mdAutenticacion.verificaToken, (req, res) => {
         usuarioAlta: req.usuario._id
     });
 
-    if (naviera.img!='' && fs.existsSync('./uploads/temp/' + naviera.img)) {
+    if (naviera.img != '' && fs.existsSync('./uploads/temp/' + naviera.img)) {
         fs.rename('./uploads/temp/' + naviera.img, './uploads/clientes/' + naviera.img, (err) => {
             if (err) { console.log(err); }
         });
     }
-    if (naviera.formatoR1!='' && fs.existsSync('./uploads/temp/' + naviera.formatoR1)) {
+    if (naviera.formatoR1 != '' && fs.existsSync('./uploads/temp/' + naviera.formatoR1)) {
         fs.rename('./uploads/temp/' + naviera.formatoR1, './uploads/clientes/' + naviera.formatoR1, (err) => {
             if (err) { console.log(err); }
         });
@@ -140,7 +140,7 @@ app.put('/:id', mdAutenticacion.verificaToken, (req, res) => {
                 errors: { message: 'No existe una naviera con ese ID' }
             });
         }
-        
+
         naviera.rfc = body.rfc;
         naviera.razonSocial = body.razonSocial;
         naviera.nombreComercial = body.nombreComercial;
@@ -158,33 +158,36 @@ app.put('/:id', mdAutenticacion.verificaToken, (req, res) => {
         naviera.caat = body.caat;
         naviera.usuarioMod = req.usuario._id;
         naviera.fMod = new Date();
-        
-        console.log (naviera);
+
         if (naviera.img != body.img) {
             if (fs.existsSync('./uploads/temp/' + body.img)) {
-                fs.unlink('./uploads/clientes/' + naviera.img, (err) => {
-                    if (err) console.log(err);
-                    else
-                        console.log('Imagen anterior fue borrada con éxito');
-                });
+                if (naviera.img != undefined || naviera.img != '' && transportista.img != null && fs.existsSync('./uploads/clientes/' + transportista.img)) {
+                    fs.unlink('./uploads/clientes/' + naviera.img, (err) => {
+                        if (err) console.log(err);
+                        else
+                            console.log('Imagen anterior fue borrada con éxito');
+                    });
+                }
                 fs.rename('./uploads/temp/' + body.img, './uploads/clientes/' + body.img, (err) => {
                     if (err) { console.log(err); }
                 });
+                naviera.img = body.img;
             }
-            naviera.img = body.img;
         }
         if (naviera.formatoR1 != body.formatoR1) {
             if (fs.existsSync('./uploads/temp/' + body.formatoR1)) {
-                fs.unlink('./uploads/clientes/' + naviera.formatoR1, (err) => {
-                    if (err) console.log(err);
-                    else
-                        console.log('Imagen anterior fue borrada con éxito');
-                });
+                if (naviera.formatoR1 != undefined || naviera.formatoR1 != '' && transportista.formatoR1 != null && fs.existsSync('./uploads/clientes/' + transportista.formatoR1)) {
+                    fs.unlink('./uploads/clientes/' + naviera.formatoR1, (err) => {
+                        if (err) console.log(err);
+                        else
+                            console.log('File anterior fue borrado con éxito');
+                    });
+                }
                 fs.rename('./uploads/temp/' + body.formatoR1, './uploads/clientes/' + body.formatoR1, (err) => {
                     if (err) { console.log(err); }
                 });
+                naviera.formatoR1 = body.formatoR1;
             }
-            naviera.formatoR1 = body.formatoR1;
         }
         naviera.save((err, navieraGuardado) => {
